use tests.BaseDatasetTest class for testing
(note: still written in Python 2.7-compatible manner)
